requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
qdrant-client==1.7.0
sentence-transformers==2.2.2
pika==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
vnstock==3.3.0
pandas==2.1.4
numpy==1.26.2

